Update button target
The class name used to target the button is not present anymore which caused the application to stop working.
It has been updated to don't use the class name anymore since it might change again in the future.

The manifest has also been updated to v3 because the v2 won't be supported from June 2023.

https://developer.chrome.com/docs/extensions/mv3/mv2-sunset/